requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto2==0.1.0
boto3==1.37.14
botocore==1.37.14
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Django==4.2.20
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
dynaconf==3.2.10
hvac==2.3.0
idna==3.10
jinja2==3.1.6
jmespath==1.0.1
knox==0.1.14
loguru==0.7.3
MarkupSafe==2.1.5
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.4
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==1.26.20
validators==0.34.0
